backend/requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
numpy
Pillow
python-multipart